pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
django = "^5.2.3"
django-environ = "^0.12.0"
pillow = "^11.2.1"
psycopg2-binary = "^2.9.10"
redis = "^6.2.0"
django-redis = "^6.0.0"
python-magic = "^0.4.27"
django-autoslug = "^1.9.9"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
